+++ aurorae/__main__.py
@@ -1,4 +1,4 @@
-from cnab.cnab240.writer import generate_cnab_files
+from aurorae.cnab240.writer import generate_cnab_files
 
 
 if __name__ == "__main__":
