+++ providers/api/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, Response
 
-from cnab.payroll.models import Payroll
+from aurorae.payroll.models import Payroll
 from providers.spreadsheet.models import Spreadsheet
