+++ providers/spreadsheet/models.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 
 from pydantic import BaseModel
 
-from cnab.cnab240.v10_7.types import CNABDate, RegistrationType, RegistrationTypeEnum
+from aurorae.cnab240.v10_7.types import CNABDate, RegistrationType, RegistrationTypeEnum
 
 
 class BaseConfig:
