+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "cnab"
+name = "aurorae"
 version = "0.1.0"
 description = "A Python implementation of the CNAB240 file to perform bulk payments."
 authors = [
@@ -67,7 +67,7 @@ mypy = "^0.910"
 freezegun = "^1.1.0"
 
 [tool.poetry.scripts]
-generate = "cnab.cnab240.writer:generate_cnab_files"
+generate = "aurorae.cnab240.writer:generate_cnab_files"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
