+++ tests/providers/spreadsheet/test_models.py
@@ -3,7 +3,7 @@ import re
 import pytest
 from pydantic.error_wrappers import ValidationError
 
-from cnab.cnab240.v10_7.types import RegistrationType, RegistrationTypeEnum
+from aurorae.cnab240.v10_7.types import RegistrationType, RegistrationTypeEnum
 from providers.spreadsheet.models import (
     Spreadsheet,
     SpreadsheetCompany,
